core/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
paho-mqtt>=1.6.1
